requirements.txt
flask==2.3.3
werkzeug==2.3.7
paramiko==3.4.0
pyftpdlib==1.5.9
pyyaml==6.0.1
pymongo==4.6.1
motor==3.3.2
python-dotenv==1.0.0
requests==2.31.0
geoip2==4.7.0
maxminddb==2.4.0
rich==13.7.0
colorama==0.4.6
discord-webhook==1.3.0
schedule==1.2.1
cryptography==41.0.7
dnspython==2.4.2
scapy==2.5.0
aiohttp==3.9.1
asyncio==3.4.3
sqlalchemy==2.0.23
python-whois==0.9.4
shodan==1.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
pyOpenSSL==23.3.0
netaddr==0.10.1